functionality merged in with customer page. Need to compute date and fix small styling details
